feat: change all type declarations of `int alpha` and `int beta` in spmv to double

+++ benchmark/acsr/ACSR.h
@@ -4,7 +4,7 @@
 #include "../utils/benchmark_time.h"
 #include "api/types.h"
 
-void acsr(int trans, const int alpha, const csr_desc<int, double> h_csr_desc, const csr_desc<int, double> d_csr_desc, const double *x,
+void acsr(int trans, const double alpha, const csr_desc<int, double> h_csr_desc, const csr_desc<int, double> d_csr_desc, const double *x,
           double *y, BenchmarkTime *bmt);
 
 #endif
