+++ benchmark/benchmark_acsr.hpp
@@ -14,7 +14,7 @@
 #include "utils/benchmark_time.h"
 
 struct ACSRSpMV : CsrSpMV {
-  void csr_spmv_impl(int trans, const int alpha, const int beta, const csr_desc<int, double> h_csr_desc,
+  void csr_spmv_impl(int trans, const double alpha, const double beta, const csr_desc<int, double> h_csr_desc,
                      const csr_desc<int, double> d_csr_desc, const double *x, double *y, BenchmarkTime *bmt) {
     acsr(trans, alpha, h_csr_desc, d_csr_desc, x, y, bmt);
   }
