+++ src/acc/hip-adaptive/adaptive.h
@@ -7,7 +7,7 @@
 
 #include "api/types.h"
 
-void adaptive_sparse_spmv(int trans, const int alpha, const int beta, const csr_desc<int, double> h_csr_desc,
+void adaptive_sparse_spmv(int trans, const double alpha, const double beta, const csr_desc<int, double> h_csr_desc,
                           const csr_desc<int, double> d_csr_desc, const double *x, double *y);
 
 #endif // SPMV_ACC_ADAPTIVE_H
