+++ src/acc/hip/spmv_hip_acc_imp.cpp
@@ -12,7 +12,7 @@
 #include "api/spmv.h"
 #include "spmv_hip_acc_imp.h"
 
-__global__ void default_device_sparse_spmv_acc(int trans, const int alpha, const int beta, int m, int n,
+__global__ void default_device_sparse_spmv_acc(int trans, const double alpha, const double beta, int m, int n,
                                                const int *rowptr, const int *colindex, const double *value,
                                                const double *x, double *y) {
   int thread_id = threadIdx.x + blockDim.x * blockIdx.x;
@@ -26,7 +26,7 @@ __global__ void default_device_sparse_spmv_acc(int trans, const int alpha, const
   }
 }
 
-void default_sparse_spmv(int trans, const int alpha, const int beta, const csr_desc<int, double> d_csr_desc,
+void default_sparse_spmv(int trans, const double alpha, const double beta, const csr_desc<int, double> d_csr_desc,
                          const double *x, double *y) {
   VAR_FROM_CSR_DESC(d_csr_desc);
   const int n = d_csr_desc.cols;
