+++ src/acc/hip/spmv_hip_acc_imp.h
@@ -10,11 +10,11 @@
 
 #include "../common/macros.h"
 
-__global__ void default_device_sparse_spmv_acc(int trans, const int alpha, const int beta, int m, int n,
+__global__ void default_device_sparse_spmv_acc(int trans, const double alpha, const double beta, int m, int n,
                                                const int *rowptr, const int *colindex, const double *value,
                                                const double *x, double *y);
 
-void default_sparse_spmv(int trans, const int alpha, const int beta, const csr_desc<int, double> d_csr_desc,
+void default_sparse_spmv(int trans, const double alpha, const double beta, const csr_desc<int, double> d_csr_desc,
                          const double *x, double *y);
 
 #endif // SPMV_ACC_SPMV_HIP_ACC_IMP_DEFAULT_H
